:bug: Fix the problem of repeated setting of execution status
