+++ composeApp/src/desktopMain/kotlin/com/clipevery/dao/task/ClipTaskRealm.kt
@@ -5,22 +5,12 @@ import org.mongodb.kbson.ObjectId
 
 class ClipTaskRealm(private val realm: Realm): ClipTaskDao {
 
-    override suspend fun update(taskId: ObjectId, block: ClipTask.() -> Unit) {
-        realm.write {
-            query(ClipTask::class, "taskId = $0", taskId).first().find()?.let {
-                return@write it.apply(block)
-            }
-        }
-    }
-
-    override suspend fun executingAndGet(taskId: ObjectId): ClipTask? {
+    override suspend fun update(taskId: ObjectId, copeFromRealm: Boolean, block: ClipTask.() -> Unit): ClipTask? {
         return realm.write {
             query(ClipTask::class, "taskId = $0", taskId).first().find()?.let {
-                it.status = TaskStatus.EXECUTING
-                it.modifyTime = System.currentTimeMillis()
-                copyFromRealm(it)
+                it.apply(block)
+                return@write if (copeFromRealm) copyFromRealm(it) else null
             }
         }
     }
-
 }
